chore: refactor methods into service
